requirements.txt
streamlit==1.51.0
anthropic==0.75.0
python-dotenv==1.2.1
